refactor and add pagination test
