+++ src/components/pagination/pagination.module.scss
@@ -1,27 +1,4 @@
-@import '../../../styles/index.scss';
-
-.container {
-  width: 100%;
-  margin: 0 auto;
-  padding-bottom: 10rem;
-}
-
-.wrapper {
-  display: flex;
-  font-family: $fontOpenSans;
-  font-size: $font_18px;
-}
-
-.universityListPage {
-  @include flex(column, center, flex-start);
-  width: 100%;
-  h1 {
-    width: 100%;
-    text-align: center;
-    font-size: $font_40px;
-    margin: 3.125rem 0 2.5rem 0;
-  }
-}
+@import '../../styles/index.scss';
 
 .pages {
   @include flex(row, center, center);
