--- src/components/superAdmin/pagination/pagination.module.scss
@@ -1,43 +1,0 @@
-@import '../../../styles/index.scss';
-
-.hoverEffect {
-  &:hover {
-    border: 2px solid $darkBlue;
-  }
-}
-
-.pagination {
-  height: 3.125rem;
-  margin-bottom: 0.625rem;
-  @include flex(row, center, center);
-
-  .leftArrow {
-    transform: rotate(180deg);
-    @extend .hoverEffect;
-  }
-
-  .rightArrow {
-    @extend .hoverEffect;
-  }
-
-  .pageNumbers {
-    margin: 0 1rem;
-    @include flex(row, center, center);
-    .number {
-      @include flex(row, center, center);
-      &:hover {
-        background-color: $darkBlue;
-        color: #fff;
-      }
-    }
-  }
-}
-
-.paginationItem {
-  cursor: pointer;
-  width: 2rem;
-  height: 2rem;
-  margin: 0.2rem;
-  border: 1px solid $gray;
-  border-radius: 0.4rem;
-}
